Exception Handling Implemented

+++ spring-tools-workspace/employee-management-app/src/com/cg/employeeapp/dao/EmployeeDaoImpl.java
@@ -5,6 +5,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import com.cg.employeeapp.exception.DuplicateEmployeeException;
+import com.cg.employeeapp.exception.EmployeeNotFoundException;
 import com.cg.employeeapp.model.Employee;
 
 public class EmployeeDaoImpl implements EmployeeDao{
@@ -16,8 +18,10 @@ public class EmployeeDaoImpl implements EmployeeDao{
 	}
 
 	@Override
-	public Employee saveEmployee(Employee employee) {
-		
+	public Employee saveEmployee(Employee employee)throws DuplicateEmployeeException {
+		if(employeeDb.containsKey(employee.getEmployeeId())) {
+			throw new DuplicateEmployeeException("Employee with ID ["+employee.getEmployeeId()+"] is already present");
+		}
 		employeeDb.put(employee.getEmployeeId(), employee);
 		return employee;
 		
@@ -30,14 +34,19 @@ public class EmployeeDaoImpl implements EmployeeDao{
 	}
 
 	@Override
-	public boolean deleteEmployeeById(int employeeId) {
-		
+	public boolean deleteEmployeeById(int employeeId)throws EmployeeNotFoundException {
+		if(!employeeDb.containsKey(employeeId)) {
+			throw new EmployeeNotFoundException("Employee with ID ["+employeeId+"] Not Found");
+		}
 		employeeDb.remove(employeeId);
 		return !employeeDb.containsKey(employeeId);
 	}
 
 	@Override
-	public Employee updateEmployee(Employee employee) {
+	public Employee updateEmployee(Employee employee) throws EmployeeNotFoundException {
+		if(!employeeDb.containsKey(employee.getEmployeeId())) {
+			throw new EmployeeNotFoundException("Employee with ID ["+employee.getEmployeeId()+"] Not Found");
+		}
 		employeeDb.put(employee.getEmployeeId(), employee);
 		return employee;		
 	}
